Organized the project in packages.
For correct usage now the user should add the project directory to
the environment variable `PYTHONPATH` so that Python knows where
to search for the executable files.

One way of doing it is this, do the following after downoading
the project:
- Go, through command line, to the root directory in which the project
was downloaded in your PC.
For example if you downloaded it to `C:\user\documents`, go to
`C:\user\documents\fasta_project`.

- Once there, you must add the path to the environment variable.
In Windows, type the followng command: `$env:PYTHONPATH = '.'`
On Linux, type: `export PYTHONPATH=.`

- Of course, the same would work if `'.'` is substituted with the absolute
path of the directory.

- This must be done each time the terminal is opened to use the project,
since the environment variable dissapears after closing it.